+++ examples/spawn.rs
@@ -161,10 +161,6 @@ fn setup(
                     .with_rotation(Quat::from_rotation_z(1.)),
                 ..Default::default()
             },
-            // Note: We don't need to manually insert an EffectProperties here, because Hanabi will
-            // take care of it on next update (since the effect has a property). Since we don't
-            // really use that property here, we don't access the EffectProperties so don't care
-            // when it's spawned. See also effect3 below for a different approach.
         ))
         .with_children(|p| {
             // Reference cube to visualize the emit origin
@@ -294,12 +290,6 @@ fn setup(
                 transform: Transform::from_translation(Vec3::new(30., 0., 0.)),
                 ..Default::default()
             },
-            // Note: We manually insert EffectProperties so update_accel() can immediately set a
-            // new value to the property, without having to deal with one-frame delays. If we let
-            // Hanabi create the component, it will do so *before* Update, so on the first frame
-            // after spawning it, update_accel() will not find it (it's spawned on next frame) and
-            // will panic. See also effect1 above.
-            EffectProperties::default(),
             DynamicRuntimeAccel,
         ))
         .with_children(|p| {
